pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdscript"
version = "0.1.0"
description = "Podcast transcription pipeline CLI"
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.0.0,<2.0.0",
  "huggingface_hub>=0.30.0,<1.0.0",
  "numpy>=1.26.0,<3.0.0",
  "openai>=1.50.0,<2.0.0",
  "pyannote.audio>=3.3.0,<4.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "soundfile>=0.12.1,<1.0.0",
  "torch>=2.2.0,<3.0.0",
  "torchaudio>=2.2.0,<3.0.0",
]

[project.scripts]
pdscript = "pdscript.cli:main"

[tool.setuptools.packages.find]
include = ["pdscript*"]
